refactor datajpatest 수정
